Added edit option on students list

+++ src/main/java/pl/lodz/uni/math/crudApplication/Student/StudentJpaRepository.java
@@ -2,6 +2,6 @@ package pl.lodz.uni.math.crudApplication.Student;
 
 import org.springframework.data.jpa.repository.JpaRepository; 
 
-public interface StudentJpaRepository  extends JpaRepository<Student, Long> {
+public interface StudentJpaRepository  extends JpaRepository<Student, Integer> {
 
 }
